test: move get_agent_score_and_set_id tests into test_agent.py for consistency with other query test files
